fix: move fixtures folder
